refactor(hover): split signature help submodule (#0000)
